// x-pack/solutions/observability/plugins/infra/common/http_api/metrics_explorer.d.ts
export declare const metricsExplorerResponseRT: rt.TypeC<{
    series: rt.ArrayC<rt.IntersectionC<[rt.TypeC<{
        id: rt.StringC;
        columns: rt.ArrayC<rt.TypeC<{
            name: rt.StringC;
            type: rt.KeyofC<{
                date: null;
                number: null;
                string: null;
            }>;
        }>>;
        rows: rt.ArrayC<rt.IntersectionC<[rt.TypeC<{
            timestamp: rt.NumberC;
        }>, rt.RecordC<rt.StringC, rt.UnionC<[rt.StringC, rt.NumberC, rt.NullC, rt.UndefinedC, rt.ArrayC<rt.ObjectC>]>>]>>;
    }>, rt.PartialC<{
        keys: rt.ArrayC<rt.StringC>;
    }>]>>;
    pageInfo: rt.TypeC<{
        total: rt.NumberC;
        afterKey: rt.UnionC<[rt.StringC, rt.NullC, rt.RecordC<rt.StringC, rt.UnionC<[rt.StringC, rt.NullC]>>]>;
    }>;
}>;